Use init script provided from upstream
Upstream packages ship init scripts that are perfectly usable nowadays
and it makes the design of this formula easier if we are using it,
rather than having to maintain various scripts for each version that
should be supported.

This patch also changes the configuration file that is being used to
/etc/influxdb/influxdb.conf which was introduced in InfluxDB 0.9.5 and
is being reflected in the init script that is being shipped in the
package.

This commit therefore breaks backwards compatibility with versions older
than 0.9.5.