Influxdb 0.12 (#13)
* Drop support for 0.8
Drop support for this influxdb version as it is severely outdated and no
longer maintained upstream.
* Use init script provided from upstream
Upstream packages ship init scripts that are perfectly usable nowadays
and it makes the design of this formula easier if we are using it,
rather than having to maintain various scripts for each version that
should be supported.
This patch also changes the configuration file that is being used to
/etc/influxdb/influxdb.conf which was introduced in InfluxDB 0.9.5 and
is being reflected in the init script that is being shipped in the
package.
This commit therefore breaks backwards compatibility with versions older
than 0.9.5.
* Move common settings into defaults.yaml
There is no point in maintaining an external os_map.yaml if the settings
therein are identical for all os flavours. This commit uses the more
common approach of defining sensible defaults in defaults.yaml that can
be optionally overridden in map.jinja.
* Install toml Python module
We plan to use this module for generating the toml configuration file
from given data, rather than having to maintain multiple templates.
* Drop support for legacy pillar structure
This commit drops support for the legacy pillar structure that was
maintained for old deployments of this formula. Some of these settings
are also no longer supported by InfluxDB itself.
* Remove obsolete logging settings
These settings have been removed in 0.9.0.
* Separate logfile settings from influxdb config
This commit separates the logfile location settings from influxdb
configuration settings that belong into its configuration file. Logfile
location is configured in the default file, rather than the influxdb
config one.
* Use standard salt file header in logrotate conf
This header makes it more obvious which file has to edited for changes
to take effect.
* Add toml_lang execution module
This module will be used for generating toml configuration files from
pillar data.
* InfluxDB default file is not version specific
The /etc/default/influxdb file is not specific to a particular InfluxDB
version and we therefore remove references to version 0.9 in its name.
* Render InfluxDB configuration from pillar data
This allows us to completely remove static configuration file templates
and generate the entire file from pillar data using the toml generator.
* Use default configuration for InfluxDB 0.9.0
The configuration file format changed drastically during the 0.9 RC
phase and this commit ensures the generation of a configuration file as
shipped in the packages for 0.9.6.1 (i.e. the latest 0.9 version).
This fixes #12
* Introduce state for InfluxDB configuration
This introduces a new influxdb.config state that will install the toml
Python module via pip and manage the InfluxDB configuration and defaults
file.It allows users to use this formula for installing InfluxDB only without
forcing them to manage the configuration file also. If they do, however,
decide to manage the configuration file the formula will behave as
expected.
* Config settings for different minor versions
This commit provides support for the definition of different
configuration defaults for minor versions.
* Add settings for InfluxDB 0.10
* Add settings for InfluxDB 0.11
* Add settings for InfluxDB 0.12
* Drop support for _latest packages
These packages are no longer supported ever since InfluxData introduced
proper package repositories on https://repos.influxdata.com
The download from http://s3.amazonaws.com/influxdb/ would fail if one
does not specify a specific version < 0.12.
* Use latest InfluxDB release from 0.12 in example